Booking views to allow the user registering a flight
We've created the booking views:
"From" view to allow the user to select start location.
"To" view to allow the user to select final location.
"Select date" view to allow the user to select date flight info.
"Passenger" view to allow the user to select how many passengers will fly.
"Results" view to show the information of the flight registered.
"Arrowback" component to allow the user going back previous screen.